# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "John Task List",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api": {},
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "description": "Create a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "John Galt"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "john.galt@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "My53cCre7P45Sw0rd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|ALrz0GZtqRLde6KlVjmWaKh1Ivn8W8WXHH3tlRIU"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Login user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "john.galt@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "My53cCre7P45Sw0rd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "3|QKIpnPo1KbBuKly4kAt23vrHHBRo7aQM38SS6A57"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "description": "Get al tasks",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Task list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create task",
                "description": "Create a new task",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/new_task_payload"
                                    }
                                ]
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "Task data",
                                        "type": "string"
                                    },
                                    "attachments[]": {
                                        "description": "Attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/new_task"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task",
                "description": "Get task details",
                "operationId": "fd3677b5b24695d7b801715869b1357c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/task"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "description": "Update task details",
                "operationId": "85377ff0e935f093dd75ea6886923dbc",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/new_task_payload"
                                    },
                                    {
                                        "properties": {
                                            "completed": {
                                                "type": "boolean",
                                                "example": "true"
                                            },
                                            "del_attachments": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "file.jpg"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "Task data",
                                        "type": "string"
                                    },
                                    "attachments[]": {
                                        "description": "Attachments",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/task"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "5ed5dfd4c645e9398f4282969bfe6478",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "new_task_payload": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "format": "text",
                        "example": "Task"
                    },
                    "description": {
                        "type": "string",
                        "format": "text",
                        "example": "Task description"
                    }
                },
                "type": "object"
            },
            "new_task": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "99695937-175d-4aae-b2b2-ea00244a7970"
                    },
                    "title": {
                        "type": "string",
                        "example": "Task"
                    },
                    "description": {
                        "type": "string",
                        "example": "Task description"
                    },
                    "author": {
                        "type": "string",
                        "example": "996b817c-9416-481f-b955-f7ab4bacaef9"
                    },
                    "attachments": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "file.jpg"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2023-06-15T23:02:09.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2023-06-15T23:02:09.000000Z"
                    }
                },
                "type": "object"
            },
            "task": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/new_task"
                    },
                    {
                        "properties": {
                            "completed": {
                                "type": "string",
                                "example": "2023-06-15T23:02:09.000000Z"
                            },
                            "editor": {
                                "type": "string",
                                "example": "996b817c-9416-481f-b955-f7ab4bacaef9"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Task ID",
                "required": true,
                "schema": {
                    "type": "string"
                },
                "examples": {
                    "uuid": {
                        "summary": "Task ID",
                        "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                    }
                }
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "",
                "scheme": "Bearer"
            }
        }
    }
}